feat: improves error handling
Creates a standard error class for API errors and adjusts the `makeRequest` utility to use it. This allows APIs to respond with an error object which will be used to populate the UI with more specific error messages and details.

Adds badges for error code and status code to the `StatusInfo` component which are shown when available (e.g. if we receive an error with code `'access_denied'` and status code `403`, two badges appear in the UI).

Replaces the three components for showing xDS configuration, envoy stats, and envoy clusters with one component because they were structurally identical.

Rearranges blocks in the StatusInfo component to rely on logical short-circuiting: when we already know the current state is “loading”, we shouldn’t let the rendering engine check again whether to display the error, empty, or data blocks. Several chained `v-if` directives are cumbersome to maintain because _all_ the underlying states need to be kept in sync (e.g. when loading is true, empty, error, and data must be false).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>