feat: reworks data planes list view
Reworks the data plane list view. The bottom portion of the view detailing the currently selected data plane was removed. As replacement, a condensed data plane summary of the selected data plane was added in the form of a fixed sidebar to the right of the data planes table. Each data plane summary includes a link to a newly added data plane detail view. The data planes table was tweaked to be more space-efficient. Only a few specific tags (kuma.io/service, kuma.io/protocol, and kuma.io/zone) are now represented in the table as separate columns. The sidebar lists all tags.

Adds a column selector to the data plane list view which allows controlling which columns are visible in the table. The selection is stored in the user browser’s local storage (i.e. its device-bound). Eventually, this selector will be added to all data tables in the UI.

Adds a data plane detail view and associated routes. It currently shows the same tabular overview that used to be shown below the data planes table in the list view.

Moves all data plane specific files into `src/data-planes`. This is motivated by wanting to structure the code base by the information hierarchy in the UI.

Creates LoadingBlock, ErrorBlock, and EmptyBlock components for simplifying the use of commonly found components (i.e. the empty states shown when loading or when content is empty or couldn’t be loaded).

Restructures the way the top level elements in the UI (i.e. the main sidebar, main content area, top bar, etc.) are laid out. This was necessary to gain a better control over the layout possibilities for the changes made to the data planes list view. As a nice side effect, a lot of unused or unnecessary code could be removed (e.g. controlling the hover state of the main navigation via CSS instead of JS).

Updates Kongponents to the latest version. This is necessary because some bug fixes made in between are necessary for the changes to the data plane views to work properly. Unfortunately, this made it necessary to regenerate all snapshots which accounts for the biggest number of files changed in this commit.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>