feat: adds service detail views
Adds detail views for service insights and external services.

Adds links to the new service insight detail views in the data plane list view.

Adds some basic logic that enables components to update the document title with more specific information (e.g. using the name of a service in the title).

Corrects a couple of issues with the mock data for services.

Adds some type annotations to API wrappers.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>