feat: simplifies main navigation
Simplifies main navigation by removing the fly out navigation that used to contain the “Service Mesh” and “Diagnostics” items. The latter was included in the regular main navigation.

Fixes not highlighting the active navigation entry when navigating to a detail view.

Turns on a setting for validating TypeScript inside Vue SFC templates and corrects a handful related issues.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>