fix: making repeated requests to zones+insights
Fixes an issue with making repeated (and often duplicated) requests to the `zones+insights/:zone` API endpoint in the data plane list view. This happened because of version incompatibility checks. Now, the the general zones overview is fetched before processing individual data planes which is generally better because data planes, if they are associated to a zone, don’t often have unique zones making the size of the response of the `zones+insights` endpoint smaller than the the added sizes of one `zones+insights/:zone` response **per data plane** (which often have the same zone).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>