refactor: moves component files into app directory
Moves all components into the app directory following a hierarchy that matches the one found in the UI. This makes it easier to find the files associated to any given page in the UI.

Updates the DEVELOPER.md file to reflect the recent changes to the project structure and organization. Most guide-like content was removed as all our tools have exhaustive documentation available that does a better job at educating developers.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>