refactor: adds BaseGuiPath variable to index file
Adds the `{{.BaseGuiPath}}` string in the index.html file generated when building for production. This variable is going to be replaced when being served by the Go application. It is expected to be replaced with the GUI base path without a leading or trailing slash (e.g. "", "gui",  "dev/gui", "a/b/c/gui", etc.).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>