refactor: reads GUI base path from DOM
Adds a `<base>` tag to the index.html file which will be set to an actual base path during development and the GO template placeholder `{{.BaseGuiPath}}` for production. Upon serving the GUI application, the placeholder will be replaced with the actual GUI base path.

Adds the `kuma-config` script tag containing `{{.}}` to the index.html file which will be replaced via Go templating with a basic data structure that, among other things, will hold the actual API base URL and the GUI base path.

Reads the API base URL and the GUI base path from the index.html file replacing any attempt of determining it from the current URL which is error-prone and inflexible.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>