refactor: cleans up styles
Cleans up a ton of things around our styles. This is done to better align the look and feel of Kuma with the Kongponents library. Besides *drastically* reducing the amount of CSS in our distribution files, this makes it easier to make use of Kongponents components because our styles don’t deviate from them (for the most part).

Throws out a ton of unused CSS (entire rule sets, unnecessary styles, no longer needed Kongponents overrides, unused custom properties and variables).

Reduces the amount of Kongponent specific overrides to a minimum.

Reduces the amount of CSS custom properties to a minimum.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>