feat(policies): Remove individual policies from the sidebar  (#503)
Removes all the Policies from the sidebar, and replaces it with one Policies link which redirects to the first Policy list with non-zero items.

Once on the policy listing, switching policies is achieved using a filter like menu which shows you only the policies of a particular type. We were too-ing and fro-ing about whether to only show policy types that have actual items in them. I've left it showing all policy items in this PR. So we end up with this kind of thing (minus the policies in the sidebar)

Signed-off-by: John Cowen <john.cowen@konghq.com>
Co-authored-by: Philipp Rudloff <philipprudloff@fastmail.com>