feat(policies): Add a new column and donut chart for policies, move starter grid down (#531)
In #503 we removed all the Policies from the side nav bar as it was getting a little busy (and that will only grow), but we noticed that it was a little weird not being able to see which PolicyTypes had non-zero counts in them at a glance.

We also noticed that the information we were after was in the [Resources] tab on the Mesh Overview page.

This PR removes the [Overview] and [Resources] tabs and combines those views, therefore when you click into the Mesh/Mesh Overview page you can immediately see which policies are populated for that Mesh. Note we don't show zero count PolicyTypes. This keeps the sidebar uncluttered but still gives you the Droidsinformation you are looking for.

Additionally we added a new DonutChart showing amount of PolicyTypes for the mesh, and lastly we moved the 'starter blocks' below all the information/data.

Signed-off-by: John Cowen <john.cowen@konghq.com>