feat(policy): Add dynamic policy route instead of generating static ones (#543)
Previously we were looping through all possible policyTypes and generating static routes for each of them.

This PR changes that to use a dynamic route for all of them instead /policies/:policyType instead. All policies now live under /policies/* instead of just /* meaning clashes are less likely.

Setting the title is a little hacky here, and setting the page title does seems to be possible via a store.dispatch to the data layer. But when I looked at setting the breadcrumb, it didn't seem possible with the current code without setting the route.meta.title within the route, ideally breadcrumbs I'd vote for breadcrumbs (and titles) being set within the *View.vue, but I didn't want to go about changing too much just for this - either via that alternative or a different one.
Signed-off-by: John Cowen <john.cowen@konghq.com>

Signed-off-by: John Cowen <john.cowen@konghq.com>