feat: adds a filter to data plane proxy tables (#505)
Adds a filter bar to data plane proxy tables which allows to filter the table for a list of known fields (i.e. `tag` and `name`).

Abstracts getting/setting query parameters into a small `QueryParameter` utility object (which replaces the `patchQueryParam` utility) to account for the need to both write _and_ read query parameters.

Adds shortcut manager utility as found in Kongponents. Preferably, the filter bar component is moved into the component library and then the utility wouldn’t be needed in this project anymore.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>