chore(docs): Move all the docs links to use env vars (#507)
This PR replaces all instances of static docs URLs with static env vars. Importantly, env vars are vars that never change without having to install a new version of Kuma, in which case you'll want to refresh your page.

I went a little further here than in the previous PR but not too far by splitting the useEnv utility/composable to use an Env service and then a very simple service container just to get us this far. The service container is likely to be switched out at some point to use something that can auto build our services for us. Important thing for me is how we use this thing, we can change how it works later.

Extra Notes:

- I changed the default version here to use 1.7.0 instead of 0.0.0 to get less snapshot churn in this PR.
- Added KUMA_CHAT_URL
- I changed a few PRODUCT_NAMEs while I was here (just the ones I came across whilst making this change). 
 Interestingly, depending on where you are in the code we seem to use different methods of getting this value. Sometimes we'd use @/constants (which is just referring to import.meta.env.VITE_NAMESPACE) and sometimes computing the value from tagline from the store, which as far as I understand is exactly the same thing. There are still some PRODUCT_NAME/tagline things in the codebase and probably a few other things that can now use useEnv/env but that can be for future PRs.
- All docs links now have UTMs
- This cleaned up some nulls in the snapshots (just thinking out loud, I wonder if theres an easy way to test for nulls in the snapshots, might make them a little more useful 🤔 )
- Added some very minimal tests for semver and useEnv
- I probably would like to change KUMA_NAME to KUMA_PRODUCT_NAME at some point, but not hugely important

Signed-off-by: John Cowen <john.cowen@konghq.com>