refactor: streamlines naming of policy type (#546)
Renames state and other variables related to policy types to distinguish the concepts of policy types (e.g. CircuitBreaker) from policies (e.g. one specific instance of a policy type).

Removes all uses of the policy type fields `singularDisplayName` and `pluralDisplayName` to be consistent with how we refer to these entities in the docs.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>