chore: Adds latest version retrieval to the mocks (#577)
This PR adds a new `KUMA_VERSION_URL` (replacing/using `VITE_VERSION_URL`) and then mocks that out using `msw`. 

This prevents us actually called out to the `VERSION_URL` during development. During tests it looks like this was originally mocked out using a similar approach (but did not read `VITE_VERSION_URL`)

I noticed that App.spec tested for no upgrade available, and we already have a test for when an upgrade is available (in `UpgradeCheck.spec`) and wanted to keep both these types of test, so I injected a super big version in `App.spec` that is higher than the currently hardcoded mocked version. I probably could have gone the other way around and mocked out the response from the endpoint also, and I'll possibly do something similar to this at a later date, this PR is already going a bit out of scope for what I want to do this for.

Probably worth noting, I'm half cheating and using `get(TOKEN......)` again here and not using native constructor/argument injection. The reason I'm doing this is to avoid all the churn which would happen if we had to reorganise all dependencies to be passed in via arguments/constructors. We should still do a re-org at a later date to completely hide the service container from the rest of the application.

Signed-off-by: John Cowen <john.cowen@konghq.com>