refactor: migrates to Chart.js
Migrates from amcharts to Chart.js. As part of this change, the labels of the individual arcs have been removed (Chart.js doesn’t have built-in support for curving the labels along an arc’s curvature) and instead, a legend beneath the chart now labels the data.

The policies chart has been removed from the mesh overview page as it only ever represented one piece of data. The total count of policies is now shown beneath next to the Policies heading.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>