chore: Adds an injectable logger (#594)
Adds an injectable logger so we can inject alternative implementations, in this case a DisabledLogger during development.

I've left the little PROD check conditional in the code the moment, just as we haven't added the development container into all our package scripts as yet. There are a couple more things I'd like to do before I make using the development container the default container during development, so I'll enable that by default in a future PR. At which point this PROD check can be removed.

Couple of extra things:

I'm not totally aware of whereabout and how we actually log in the code, but on a brief scan, its look like in the future we could potentially use our logger directly instead of accessing datadogLogs.logger directly. One for the future tho.
I've added some README in here which is empty for the moment, but is likely to be filled out in the future. I know this is a little redundant right now, but I would rather each 'service' have a README, even if its empty to begin with.

Signed-off-by: John Cowen <john.cowen@konghq.com>