chore: Adds injectable API client (#602)
This PR moves our `kumaApi` client to be an injectable/injected object.

In order to do this I didn't want to provide a `useAPI` or similar and go through the entire codebase changing everything as that would take quite a bit of churn. So I replaced the instantiation of `kumaApi` in `@/api/kumaApi` with a reference to the new API service in the service container.

As the `store` files where importing `kumaApi` and we are already importing `store` in the service container/composition root, this meant we have to avoid the circular dependencies. So I made some minimal tweaks to store so that kumaApi could be injected there also, and used a similar approach where we instantiate store so that the entire application uses the store from the service container.

I've also moved all the kumaApi related files/class into its own service folder in `@/services/kuma-api` (a lot of the green in the changset is this), but I've stopped short doing this for store related files for the moment at least.

Signed-off-by: John Cowen <john.cowen@konghq.com>