fix: looking up external service with wrong name
Fixes a bug where in the service list view, an external service resource was looked up using the service insight name which isn’t correct. For now, in the absence of the lookup ability via API, we find the corresponding external service in the list of all external services.

Fixes #599.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>