Merge branch 'release-2.1' into chore/backport-changes-from-release-2.1-branch (#616)
* fix: looking up external service with wrong name (#607)

* fix: looking up external service with wrong name

Fixes a bug where in the service list view, an external service resource was looked up using the service insight name which isn’t correct. For now, in the absence of the lookup ability via API, we find the corresponding external service in the list of all external services.

Fixes #599.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* ci: creates unique branch name (backport) (#610)

ci: creates unique branch name (#609)

Uses a branch name that’s unique between the target branches (e.g. when merging into a release branch).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix: reading subscriptions too early (#611)

Fixes an issue of accessing a zone insight’s subscriptions too early when briefly after creating a zone, the zoneInsight object is not available, yet.

Fixes #585.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>