chore: Makes App injectable, moves injections to composition root/main (#644)
This commit re-arranges initializeVue slightly to no longer use get(TOKENS.*) and use parameter/constructor injection instead. This is partly to further hide our service container from the actual application code, partly to move the injections to the very root of the application (main.ts) and partly to allow us to inject App if we ever need to make use of that. We continue to do the majority of our application wiring in @/services and import that into main so we can get the 'application' and call it.

I also added "main": "index.ts" to the package.json meaning kuma-gui can now be used as a real dependency/sub-package.

Signed-off-by: John Cowen <john.cowen@konghq.com>