fix: base API URL being set incorrectly
Fixes an issue with the base API URL being set incorrectly. To fix this issue, the logic for ensuring the API base URL is indeed an absolute URL and not just a path segment relative to the origin was moved into the Env.ts file where the Kuma config object is being fetched. `RestClient` now always expects an absolute URL without any trailing slashes.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>