feat: Accept "string only" versions (#698)
Loosens up our version related things so we can use string based versions if we need to. The rule it uses is "if major doesn't parse as a number, then its a 'string version'". String versions will never ask for updates as they are more likely to be dev versions and suchlike.

I switched a couple of old store-type things to use env vars whilst I was here.

The entire thing assumes that we never deal with actual full semvers, only 3 numbers i.e. we never need to be aware of carets and tildes, only 3 numbers joined by dots `~3.0.0` vs `3.0.0`.

This PR makes it so we only need to be aware of "3 numbers joined by dots" and "full strings" - we still don't need to be aware of `~` or `^` prefixed semvers.

Signed-off-by: John Cowen <john.cowen@konghq.com>