feat: enables mock worker to handle multiple api instances (#719)
Changes `setupMockWorker` to take `apiName` and `handlers` parameters. The function now keeps a record of which handlers belong to which API (as identified by `apiName`). This way, we can store independent handlers per API which might need to be constructed in different ways.

Makes sure to stop the mock worker before starting it again which seems to be necessary despite a logged warning indicating the opposite. Without it, newly added handlers have no effect.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>