refactor: makes RestClient injectable
Reduces `Api` base class to most minimal abstraction for API instances to inherit from. Its only purpose is to provide the constructor for setting `client` and `Env` so that API instances don’t have to do that. No further functionality is provided. This is now the sole responsibility of `RestClient`.

Makes `RestClient` a constructor-provided parameter of API instances. This allows different environments to construct API instances with different `RestClient` instances. For example, `KumaApi` is now constructed with a plain `RestClient` instances in the production environment. In the development environment, it is however provided a mocked variant via the new `createMockedRestClient`.

Moves mocking responsibilities from individual API instances to `RestClient`. Specifically, a new `createMockedRestClient` factory function is provided which takes a set of `mocks` and the API instance-specific `setupHandlers` function as arguments.

Sets the `credentials` option to default to `'include'` for all instances of `RestClient`. Environments in which no session state is transferred via cookies aren’t affected as there wouldn’t be any cookie stored to send with requests. Environments which do need to transfer such session state in turn will benefit from not needing to set `credentials: 'include'` on all the individual API instances.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>