fix: not bundling all font files
Fixes a bug with the build not bundling all font files which was caused by the extracted woff2 files actually being _identical_ between different weights which Rollup actually seems to detect and optimize away. Deleting the redundant files and using the same file name for all `@font-face` rules seems to fix the issue.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>