chore: Root all API URLs (#785)
We mostly hardcode all our URLs and hide them within a 'data retrieval' class. When I say mostly, we have complete hardcode-able control over them, but they are made up of dynamic segments. As we have complete control over them, we can hardcode every one of them to be 'rooted'. i.e. `/meshes/${name}` vs `meshes/${name}`

We also have almost complete control over the baseURL for all our API requests from our application. This means we can ensure that our baseURL does not have a trailing slash i.e. `http://localhost:5681` vs `http://localhost:5681/` (it's already like this in our development build, so no change for this in this PR)

This means that if we stick to the "rooted URLs with a non trailing slash baseURL" rule. In order to concatenate URLs we just do:

```
${baseURL}${path}
```

...which removes any complications, plus the need for any tests that go with testing any of that complication, and any code scanning issues in relation to the regexs we were using. 

Notes:
- We continue to use a prefix of `http` to decide whether something requires a `baseURL` adding or not.
- We 'almost' have complete control over our baseURL. During development we control this value via a hardcoded env var (which currently has no trailing slash), but in production it is passed to us via the backend. I'm currently unsure whether this is validated/processed to ensure it has no trailing slashes before being passed to us. Ideally I would slightly prefer it to be done in the backend, but if we can't do that for whatever reason, I can add a `stripLastSlashes(baseURL)` on the frontend instead (possibly a post-review follow up commit on this PR).
- There's probably a few more RestClient tests I could remove here as they aren't really testing anything apart from `${baseURL}${path}` which isn't really worth testing.

We decided to keep the URL sanitizing/normalizing in the frontend, we've put it as close to where it comes in from the backend as possible.


Signed-off-by: John Cowen <john.cowen@konghq.com>