chore: tweaks kcard-list usage (#768)
* chore: tweaks kcard-list usage

Renames `kcard-list` to `kcard-switcher` and tweaks its threshold to be more applicable to our use cases (e.g. for wrapping the mesh resources items).

Adds a new `columns` class for laying out a list of items using a preferred number of columns (3 by default). These items will wrap one by one (as opposed to all at the same time with `kcard-switcher`) when their space requirements are no longer met.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix(zones): showing empty entry

Fixes showing an empty entry in the zone ingress insights if its data is `{}` (this manifested in showing an empty "rds" entry using mock data).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: regenerates snapshots

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor: uses AppGap property for consistent gaps

Introduces a global `--AppGap` property (with value `--spacing-lg`) and uses it for gaps between cards and general "top-level" spacing. This provides a central place to configure this kind of spacing. This changes the spaces between cards from `--spacing-md` to `--spacing-lg`.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>