feat: teaches breadcrumbs to read chains of parent routes
Teaches the breadcrumbs component to read a complete chain of parent routes instead of only ever just one route-parent route relationship. This allows us to display, for example, "Policies → CircuitBreaker → cb-1".

Ensures that the title of a parent route is constructed in the same way as the one for the current route (i.e. taking `getBreadcrumbTitle` and `breadcrumbTitleParam` into account).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>