fix: determining wrong policy type route to redirect to
Makes sure that a policy type for which no property exists on `store.state.sidebar.insights.mesh.policies` isn’t actually returned as the policy type to redirect to.

Also makes sure that, in the unlikely case that the policies redirect is triggered before `store.state.policyTypes` is populated, no error happens on account of accessing `item.path`: the redirect will then go to the home route. Technically, application logic should ensure that the policy types are present at that point, so this is just a safety measure.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>