fix: rendering unwanted components before routing is done
Fixes an issue where some components in `App.vue` would render before we could reliably determine the state that actually tells us whether we want to render them. That happens because when navigating, vue-router initially (on first load) load a `/` route with name `undefined`. When that happens, properties like `route.meta.isWizard` won’t be set matching the actually targetted route. This is already undesirable but it also leads to a bug where `GithubButton` falls out of the render cycle too quickly and throws an error.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>