feat: adds Zone creation flow
Adds the ZoneUpdateView component and the corresponding routes and controls in the zone list view for creating and deleting zones. The functionality for accessing the zone creation flow is enabled using the `VITE_ZONE_CREATION_FLOW` environment variable which, for now, is only enabled in development mode (like a feature flag).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>