feat: adds Zone creation flow (#869)
* feat: adds WizardTitleBar and FormProgressBar components

Adds dedicated WizardTitleBar and FormProgressBar components for future re-use in other wizards.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor: changes top-level layout for wizards

Changes the top-level layout so that new style wizards can claim all space and not show either sidebar or top bar.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix: replaces < and > before syntax highlighting

Fixes a bug with the syntax highlighting removing content looking like HTML tags (e.g. `<some string>`).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: expands the mocks around zones a little bit

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix: rendering unwanted components before routing is done

Fixes an issue where some components in `App.vue` would render before we could reliably determine the state that actually tells us whether we want to render them. That happens because when navigating, vue-router initially (on first load) load a `/` route with name `undefined`. When that happens, properties like `route.meta.isWizard` won’t be set matching the actually targetted route. This is already undesirable but it also leads to a bug where `GithubButton` falls out of the render cycle too quickly and throws an error.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor(e2e): allows inferring $ type

Rewrites `$` to allow inferring its type correctly as with the current implementation, `cy.get` lost its generic-ness and would lead to type errors in legitimate code.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor(e2e): simplifies step utilities

Refactors the step for clicking interactive UI elements so that it avoids the use of `{ force: true }` by attempting to click the associated label for an input element.

Adds a step utility for asserting certain boolean HTML attributes (e.g. `disabled`, `required`, `readonly`).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* test(e2e): adds create Zone flow test

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* feat: adds i18n.t function to useI18n utility

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* feat: adds Zone creation flow

Adds the ZoneUpdateView component and the corresponding routes and controls in the zone list view for creating and deleting zones. The functionality for accessing the zone creation flow is enabled using the `VITE_ZONE_CREATION_FLOW` environment variable which, for now, is only enabled in development mode (like a feature flag).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix: entity scanner not awaiting props.loaderFunction

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore(e2e): removes "do" and "does"

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore(e2e): removes "check" option

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>