feat: adds mesh list view (#898)
* chore: updates mesh-insights mock

Updates the mesh-insights mock to have a bunch more random status values.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore(types): matches MeshInsight close to backend

Updates the `MeshInsight` type and some related types to be close to the backend.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor: nests mesh routes to match usual structure

Splits the `/mesh/:mesh` route record into a `/mesh` record with only one children route record which is the `:mesh` route record which allows for adding routes _next_ to `:mesh` without inheriting the `mesh` route parameter (e.g. for the mesh list view route).

Adds the Meshes breadcrumbs item when on a page under or directly on `/mesh`. Adds the current mesh (e.g. default) when on a page under `/mesh`.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* feat: adds mesh list view

Adds a MeshListView component at `/mesh`.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>