chore: YAML for i18n layer (#913)
Adds a YAML layer for our i18n strings.

---

I've also added the ability to tag YAML text as being markdown formatted, which may make things easier for folks in the future. We don't use this as yet but I'm guessing I will use it in the future, so I've left it in here for now (unless there are any concerns with doing that)

See MADR for more details

Additionally I've added a couple of amends to the Kuma YAML configs contained in the YAML i18n strings here. That have been spoken about previously. I'll inline comment them so its clear where they are.

I also noticed that having camelCase i18n property names can be problematic (I've added both forms of mTLS that we use in the API), so I would recommend lowercase i18n property names where possible, and I'm considering toLowerCase()ing all our paths that are passed to t at some point in the future to prevent this.

`jest` yaml loader required until we get rid of `jest`, at which point it can be deleted.

Signed-off-by: John Cowen <john.cowen@konghq.com>