feat: always shows resources in universal format
Reworks the `YamlView` component into a more independent `ResourceCodeBlock` component with the built it ability to fetch the right resource via the API: instead of getting the resource as an object, it now gets a wrapper around the appropriate API method to retrieve the resource by itself. The component also no longer offers the two display formats Universal and Kubernetes in the form of tabs but rather always shows Universal with the option to copy the Kubernetes format.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>