feat: moves zones into tabs
Renames the routes associated to the Zone CP views to include the `-cp` bit to make it more clear this is a view of Zone CPs.

Moves the Zone sidebar items into a Zones page with navigational tabs which shows the Zone CPs view by default (i.e. like we’ve done for mesh-related items).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>