feat: change wording of status not available (#1026)
* fix: not showing status in services list view

Fixes not showing a service’s status in the service list view at all.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* feat: changes wording for "not_available" status

Changes the wording for the `not_available` status from "Not available" to "Information not available". (I’ve added both camel-cased and snake-cased variants to the messages file as we sometimes use one, sometimes the other. If we change this, we can drop one.)

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor: switch from property to value

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor: simplifies status badge class names

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: removes unused shouldHideTitle prop

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: regenerates snapshots

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>