chore: View layer based titling and breadcrumbing (#965)
Removes titling and breadcrumbing out of the routing layer and into the view layer.

- Every "Route/View Component" should be wrapped in a `RouteView` (similar to `RouterView` being a 'marker' for a child route)
- Every "application page" should be wrapped in an `AppView`, here you can add breadcrumbs or not (or specify that you shouldn't have breadcrumbs at all with `[]`)
- Anything that amends the title should use `AppTitle`, which along with `RouteView` provides HTML titling and a11y route announcing.

Overall this means that titling and breadcrumbing are no longer tied to the top parent App.vue in a large nested routing structure, and means that App.vue doesn't necessarily need to be at the absolute top level of the routing structure in order to be able to add titles and breadcrumbs. It's almost an inheritance vs composition shift. Whilst this is an advantage none of this work exists in this commit.

Signed-off-by: John Cowen <john.cowen@konghq.com>